Create a function giveScore
giveScore() function gives score to the player or the computer according
to the result of every round.

If the letter at the 13th index of resultOfTheRound variable is 'W',
that means player win the round, then giveScore() will increment the
playerScore.

Else if the letter at the 13th index of resultOfTheRound variable is 'L'
that means computer win the round, then giveScore() will increment the
computerScore.